move containers into container folder
