+++ client/src/index.tsx
@@ -1,8 +1,10 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import './index.css';
-import App from './App';
+
+import App from './containers/App';
 import * as serviceWorker from './serviceWorker';
+import './css/normalize.css';
+import './css/index.css';
 
 ReactDOM.render(<App />, document.getElementById('root'));
